Backend/requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
# psycopg2-binary
Flask-WTF
Flask-CORS
Flask-APScheduler
Flask-Mail
Flask-Migrate
python-dotenv